requirements/local.txt
# Local development dependencies go here
-r base.txt

coverage==4.4.1
django-coverage-plugin==1.5.0

Sphinx==1.6.2
django-extensions==1.7.9
Werkzeug==0.12.2
django-test-plus==1.0.17
factory-boy==2.8.1

django-debug-toolbar==1.8

# improved REPL
ipdb==0.10.3

pytest-django==3.1.2
pytest-sugar==0.8.0

pandas==0.19.2
